# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tasty Point Restaurant QR Ordering API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderPaymentStatus } from "./orderPaymentStatus";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  tableId: string;
  tableNumber: number;
  status: OrderStatus;
  items: OrderItem[];
  totalAmount: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  specialRequests?: string | null;
  paymentStatus?: OrderPaymentStatus;
  /** @nullable */
  razorpayOrderId?: string | null;
  /** @nullable */
  razorpayPaymentId?: string | null;
  createdAt: string;
}