# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tasty Point Restaurant QR Ordering API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputItemsItem } from "./orderInputItemsItem";

export interface OrderInput {
  tableId: string;
  customerName?: string;
  specialRequests?: string;
  items: OrderInputItemsItem[];
}