# lib/api-zod/src/generated/types/paymentVerifyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tasty Point Restaurant QR Ordering API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentVerifyInput {
  orderId: string;
  razorpayOrderId: string;
  razorpayPaymentId: string;
  razorpaySignature: string;
}